Scraper/pincode_stock_checker/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
twilio>=8.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
firebase-admin>=6.0.0
psycopg2-binary>=2.9.0
